kuka_iiwa: Move helpers from examples to manipulation
