+++ examples/kuka_iiwa_arm/BUILD.bazel
@@ -36,6 +36,7 @@ drake_cc_library(
         "//attic/multibody/parsers",
         "//common:find_resource",
         "//common/trajectories:piecewise_polynomial",
+        "//manipulation/kuka_iiwa:iiwa_constants",
         "@lcmtypes_bot2_core",
         "@lcmtypes_robotlocomotion",
     ],
@@ -54,6 +55,7 @@ drake_cc_library(
         "//attic/multibody/rigid_body_plant:contact_results",
         "//common/trajectories:piecewise_polynomial",
         "//lcmtypes:iiwa",
+        "//manipulation/kuka_iiwa",
         "//systems/framework:leaf_system",
         "//systems/lcm:lcm_pubsub_system",
         "//systems/lcm:translator",
