+++ examples/kuka_iiwa_arm/iiwa_common.cc
@@ -29,21 +29,6 @@ namespace drake {
 namespace examples {
 namespace kuka_iiwa_arm {
 
-VectorX<double> get_iiwa_max_joint_velocities() {
-  // These are the maximum joint velocities given in Section 4.3.2 "Axis data,
-  // LBR iiwa 14 R820" of the "LBR iiwa 7 R800, LBR iiwa 14 R820 Specification".
-  // That document is available here:
-  // https://www.kuka.com/-/media/kuka-downloads/imported/48ec812b1b2947898ac2598aff70abc0/spez_lbr_iiwa_en.pdf
-  return (VectorX<double>(7) << 1.483529,  //  85°/s in rad/s
-          1.483529,                        //  85°/s in rad/s
-          1.745329,                        // 100°/s in rad/s
-          1.308996,                        //  75°/s in rad/s
-          2.268928,                        // 130°/s in rad/s
-          2.356194,                        // 135°/s in rad/s
-          2.356194)                        // 135°/s in rad/s
-      .finished();
-}
-
 template <typename T>
 Matrix6<T> ComputeLumpedGripperInertiaInEndEffectorFrame(
     const RigidBodyTree<T>& world_tree,
